+++ lib/rspec/mocks/framework.rb
@@ -4,7 +4,7 @@
 
 require 'rspec/mocks/configuration'
 require 'rspec/mocks/extensions/instance_exec'
-require 'rspec/mocks/stashed_instance_method'
+require 'rspec/mocks/instance_method_stasher'
 require 'rspec/mocks/method_double'
 require 'rspec/mocks/methods'
 require 'rspec/mocks/argument_matchers'