+++ spec/rspec/mocks/instance_method_stasher_spec.rb
@@ -2,7 +2,7 @@ require 'spec_helper'
 
 module RSpec
   module Mocks
-    describe StashedInstanceMethod do
+    describe InstanceMethodStasher do
       class ExampleClass
         def hello
           :hello_defined_on_class
@@ -17,7 +17,7 @@ module RSpec
         obj = Object.new
         def obj.hello; :hello_defined_on_singleton_class; end;
 
-        stashed_method = StashedInstanceMethod.new(singleton_class_for(obj), :hello)
+        stashed_method = InstanceMethodStasher.new(singleton_class_for(obj), :hello)
         stashed_method.stash
 
         def obj.hello; :overridden_hello; end
@@ -32,7 +32,7 @@ module RSpec
         def obj.hello; :hello_defined_on_singleton_class; end;
         singleton_class_for(obj).__send__(:private, :hello)
 
-        stashed_method = StashedInstanceMethod.new(singleton_class_for(obj), :hello)
+        stashed_method = InstanceMethodStasher.new(singleton_class_for(obj), :hello)
         stashed_method.stash
 
         def obj.hello; :overridden_hello; end
@@ -43,7 +43,7 @@ module RSpec
       it "only stashes methods directly defined on the given class, not its ancestors" do
         obj = ExampleClass.new
 
-        stashed_method = StashedInstanceMethod.new(singleton_class_for(obj), :hello)
+        stashed_method = InstanceMethodStasher.new(singleton_class_for(obj), :hello)
         stashed_method.stash
 
         def obj.hello; :overridden_hello; end;